Suggested fix for theta_offset problem in PolarizedNeutronProbe
The optimization of just adding a fixed offset when all the theta_offsets are shared
in a PNP object does not work, particularly if the data needs to be oversampled.

This commit removes the optimized version which doesn't work, and instead adds an
"oversampling" argument to the init of PolarizedNeutronProbe.

If the theta_offsets change (if they are a fit parameter) the measurement_union and
the resampling are recalculated.  This is slow but there may be no way around this
without a mesh of theta/lambda values to interpolate from.